Add: Convert app to React Router
For all features and functions needed for this application I chose to
use React router. Initially I set up this application with a basic react
app, but now that it's ready to have the backend OAuth connected to the
front end the app needed to be refactored to use React router.