Add: delay utility function to prevent db error on running seed file
When running the seed script the_odin_book database would be dropped
and then recreated, but the local mysql db running on this machine
would not have recognised this creation yet so I when trying to swtich
to the newly create the_odin_book db it would throw an error starting
that the db did not exist yet. By adding this function and forcing a 2
second pause between creating the db and sending db queries this bug
was resolved.