Add: status modal styles and func to send post req to '/api/posts/'
Added a modal that open when the status input button is clicked.
Once open the user has the abillity to enter text into a textarea and
then sumbit the post via an axio post request to the '/api/posts/ route.
The user also has the option to click the X in the top right of the
modal and this will close the modal.