Add: createdNow boolean paramater for time stamping
Now that I have moved out of the database seeding part of development
I am adding new data to the db from the application in real time. When
creating a user, post, commnet, etc. I wanted to ability to use a random
time stamp or the current data time. The addition of the createdNow
boolean provides that functionality.