Add: post like handling and db updating via axios
Implemented like tracking via boolean state variable that is toggled
when a user click on the post like button. When the state is updated
a submitPostLike() is called which hits the server with an axios post
request and postLike payload. If postLike state is truthy then postLike
is inserted into DB if it is falsy then the postLike matching both
userID and postID is deleted from postLike db table.

Additionaly I added an isMounted state variable to track if this is the
first rendering of the component to prevent the useEffect hook from
calling the submitPostLike() function on initial render.

Note: Hitting the database on evey use click of the like button seems
inefficient at best thus I plan to revist this component to improve
performance, but for the sake of moving the project along
I will delay this refactoring to another time.