Add: friendship db seed function
This function adds a random number of friendships to each user in the
db. It also makes sure that repeate friendships do not occur and the a
user is not friended with themsleves.

Due to the looped nature of these checks and corrections that needed to
be made to ensure the above I needed to store all the correct friendship
data into one large array and then batch insert this via the mysql2
library.